requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2

# Deep Learning (CPU-only version)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0+cpu
torchvision==0.16.0+cpu
numpy>=1.26,<2.0
pillow==10.1.0

# Data Processing
pandas==2.1.3
scikit-learn==1.3.2

# Monitoring & Logging
prometheus-client==0.19.0
mlflow==2.8.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.1

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
pylint==3.0.2
mypy==1.7.1
bandit==1.7.5

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

# Data Versioning (Optional)
dvc[s3]==3.30.1

# Additional
typing-extensions==4.8.0